Garmin protocol: change naming and logic of several FIT classes
- refactor the logic of Global and Local messages
- add some Global messages with naming taken from [1]
- Global messages are not enum because there are too many
- introduce the concept of FieldDefinitionPrimitive
- add new Field Definitions
- add support for developer fields and array fields
- add test case for FIT files taken from [0]

[0] https://github.com/polyvertex/fitdecode/
[1] https://www.fitfileviewer.com/

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/garmin/fit/baseTypes/BaseTypeInt.java
@@ -32,7 +32,7 @@ public class BaseTypeInt implements BaseTypeInterface {
             return null;
         if (i == invalid)
             return null;
-        return (int) ((i + offset) / scale);
+        return ((i + offset) / scale);
     }
 
     @Override
